backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-dotenv==1.0.1
anthropic==0.28.0
python-docx==1.1.2
weasyprint==62.3
jinja2==3.1.4
httpx==0.27.0
pydantic==2.7.0
pillow==10.3.0
